+++ frontend/src/pages/ProjectDetails/Sidebar/View.tsx
@@ -1,7 +1,7 @@
 import { Listbox } from "@headlessui/react";
 import UpDownChevrons from "src/assets/icons/UpDownChevrons";
 import { RoutePaths } from "src/App";
-import BackLink from "src/components/BackLink";
+import BackLink from "src/pages/ProjectDetails/Sidebar/BackLink";
 import RoundedImage, { ImageSize } from "src/components/RoundedImage";
 import { ProjectDetails } from "..";
 import { useIntl } from "src/hooks/useIntl";
